src/set-impl.h: Zero-out structs allocated from the pool.
In macros SET_ALLOCATE_STRUCT and SET_ALLOCATE_VAR_STRUCT,
honour the contract of zeroing-out the structs allocated
from the memory pool. This amounts to an embarrasing bug,
which did not surfaced out yet because the memory pool is
of fixed size preallocated via 'malloc'. Glibc's 'malloc'
is returning zeroed-out blocks when allocating them the
first time from the OS (a GNU/Linux platform) -- the OS
itself gave the blocks back to 'malloc' as such.